requirements.txt
pillow==10.4.0
requests==2.32.3
imageio==2.34.2
numpy==2.0.0